fixie-agent/pyproject.toml
[tool.poetry]
name = "mimo-agent"
version = "0.1.0"
description = ""
authors = ["Mohamed Elattma <mo@mimo.team>"]
readme = "README.md"
packages = [{include = "mimo_agent"}]

[tool.poetry.dependencies]
python = "^3.10"
fixieai = "^0.2.10"
requests = "^2.28.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
